+++ recipes/tasks.py
@@ -3,7 +3,6 @@
 import shlex
 import subprocess  # nosec # noqa S404
 import sys
-from typing import List
 
 from beartype import beartype
 from calcipy.doit_tasks.base import debug_task
@@ -80,7 +79,7 @@ def task_compress() -> DoitTask:
         DoitTask: DoIt task
 
     """
-    def _run_params(pos: List[str]) -> None:
+    def _run_params(pos: list[str]) -> None:
         for pos_arg in pos:
             cmds = shlex.split(f'{_OPTIMIZE_CMD} {pos_arg}')
             logger.info('Running: {cmds}', cmds=cmds)